Removed one level of inheritance from the controllers
We didn't need the extra level of inheritance since the implementation of controllers aren't in a separate assembly
