--- src/Hero6.Engine/ModuleController/Campaigns/Characters/MonoGameCharacterController.cs
@@ -1,211 +1,0 @@
-﻿// <copyright file="MonoGameCharacterController.cs" company="Late Start Studio">
-// Copyright (C) Late Start Studio
-// This file is subject to the terms and conditions of the MIT license specified in the file
-// 'LICENSE.CODE.md', which is a part of this source code package.
-// </copyright>
-
-using System.Collections.Generic;
-using System.Linq;
-using LateStartStudio.Hero6.Extensions;
-using LateStartStudio.Hero6.ModuleController.Campaigns.Animations;
-using LateStartStudio.Hero6.ModuleController.Campaigns.Characters.InventoryItems;
-using LateStartStudio.Hero6.ModuleController.Campaigns.Characters.Stats;
-using LateStartStudio.Hero6.ModuleController.Campaigns.InventoryItems;
-using LateStartStudio.Hero6.ModuleController.Campaigns.Rooms;
-using LateStartStudio.Hero6.ModuleController.Campaigns.Rooms.Regions;
-using LateStartStudio.Hero6.Services.Campaigns;
-using LateStartStudio.Hero6.Services.DependencyInjection;
-using Microsoft.Xna.Framework;
-
-namespace LateStartStudio.Hero6.ModuleController.Campaigns.Characters
-{
-    public class MonoGameCharacterController : CharacterController
-    {
-        private readonly ICampaigns campaigns;
-        private readonly MonoGameStatsController stats;
-        private readonly List<MonoGameInventoryItemController> inventory = new List<MonoGameInventoryItemController>();
-        private readonly Queue<Point> path = new Queue<Point>();
-
-        private double pixelsToMove;
-        private MonoGameCharacterAnimationController currentAnimation;
-        private MonoGameCharacterAnimationController idleAnimation;
-        private MonoGameCharacterAnimationController moveAnimation;
-
-        public MonoGameCharacterController(ICharacterModule module, IServiceLocator services)
-            : base(module, services)
-        {
-            campaigns = services.Get<ICampaigns>();
-            stats = new MonoGameStatsController(services);
-        }
-
-        public override int Width => currentAnimation?.Width ?? 0;
-
-        public override int Height => currentAnimation?.Height ?? 0;
-
-        public override bool IsPlayer => campaigns.Current.Player == Module;
-
-        public override int Speed { get; set; } = 60;
-
-        public override IRoomController Room
-        {
-            get
-            {
-                return campaigns.AsMonoGame().CurrentController.Rooms.Values.Single(r => r.Characters.Contains(this));
-            }
-        }
-
-        public override IStatsController Stats => stats;
-
-        public override ICharacterAnimationController IdleAnimation
-        {
-            get { return idleAnimation; }
-            set { idleAnimation = campaigns.AsMonoGame().CurrentController.CharacterAnimations[value.Module.GetType()]; }
-        }
-
-        public override ICharacterAnimationController MoveAnimation
-        {
-            get { return moveAnimation; }
-            set { moveAnimation = campaigns.AsMonoGame().CurrentController.CharacterAnimations[value.Module.GetType()]; }
-        }
-
-        public override IEnumerable<IInventoryItemController> Inventory => inventory;
-
-        public override void AddInventoryItem<T>()
-        {
-            inventory.Add(campaigns.AsMonoGame().CurrentController.InventoryItems[typeof(T)]);
-        }
-
-        public override void RemoveInventoryItem<T>()
-        {
-            inventory.Remove(campaigns.AsMonoGame().CurrentController.InventoryItems[typeof(T)]);
-        }
-
-        public override bool HasInventoryItem<T>()
-        {
-            return inventory.Contains(campaigns.AsMonoGame().CurrentController.InventoryItems[typeof(T)]);
-        }
-
-        public override void Walk(int x, int y)
-        {
-            var from = new Point(X, Y);
-            var to = new Point(x, y);
-            var newPath = ((MonoGameWalkAreasController)Room.WalkAreas).GetPath(from, to);
-
-            if (newPath == null)
-            {
-                return;
-            }
-
-            path.Clear();
-            newPath.ToList().ForEach(p => path.Enqueue(p));
-            path.Dequeue(); // First is where we're already standing so discard
-        }
-
-        public override void Face(CharacterDirection direction)
-        {
-            if (idleAnimation != null)
-            {
-                idleAnimation.Direction = direction;
-            }
-
-            if (moveAnimation != null)
-            {
-                moveAnimation.Direction = direction;
-            }
-        }
-
-        public override void Face(int locationX, int locationY)
-        {
-            var from = new Point(X, Y);
-            var to = new Point(locationX, locationY);
-            var direction = to - from;
-            Face(direction.ToVector2().ToCharacterDirection());
-        }
-
-        public override void ChangeRoom<T>(int x = 0, int y = 0, CharacterDirection direction = CharacterDirection.CenterDown)
-        {
-            path.Clear();
-            X = x;
-            Y = y;
-            Face(direction);
-            Room.RemoveCharacter(Module);
-            campaigns.AsMonoGame().CurrentController.Rooms[typeof(T)].AsMonoGame().AddCharacter(this);
-        }
-
-        public override void SetAsPlayer() => campaigns.Current.Player = Module;
-
-        public override bool Interact(int x, int y, Interaction interaction)
-        {
-            if (currentAnimation.Interact(x, y, interaction))
-            {
-                switch (interaction)
-                {
-                    case Interaction.Eye:
-                        Module.Look?.Invoke();
-                        break;
-                    case Interaction.Mouth:
-                        Module.Talk?.Invoke();
-                        break;
-                    case Interaction.Hand:
-                        Module.Grab?.Invoke();
-                        break;
-                    case Interaction.Move:
-                    default:
-                        return false;
-                }
-
-                return true;
-            }
-
-            return false;
-        }
-
-        public override void Load()
-        {
-        }
-
-        public override void Unload()
-        {
-        }
-
-        public override void Update(GameTime time)
-        {
-            Move(time);
-            currentAnimation = path.Count > 0 ? moveAnimation : idleAnimation;
-            currentAnimation.X = X;
-            currentAnimation.Y = Y;
-            currentAnimation.Update(time);
-        }
-
-        public override void Draw(GameTime time)
-        {
-            currentAnimation.Draw(time);
-        }
-
-        private void Move(GameTime time)
-        {
-            if (path.Count == 0)
-            {
-                return;
-            }
-
-            pixelsToMove += Speed * time.ElapsedGameTime.TotalSeconds;
-            var pixelsToMoveFloor = (int)pixelsToMove;
-            pixelsToMove -= pixelsToMoveFloor;
-
-            for (var i = 0; i < pixelsToMoveFloor; i++)
-            {
-                var newLocation = path.Dequeue();
-                Face(newLocation.X, newLocation.Y);
-                X = newLocation.X;
-                Y = newLocation.Y;
-
-                if (path.Count == 0)
-                {
-                    pixelsToMove = 0.0;
-                    break;
-                }
-            }
-        }
-    }
-}
