+++ src/Hero6.Engine/ModuleController/Campaigns/Characters/Stats/StatsController.cs
@@ -4,116 +4,106 @@
 // 'LICENSE.CODE.md', which is a part of this source code package.
 // </copyright>
 
+using LateStartStudio.Hero6.Services.Campaigns;
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using Microsoft.Xna.Framework;
 
 namespace LateStartStudio.Hero6.ModuleController.Campaigns.Characters.Stats
 {
     /// <summary>
     /// API for Stats Controller.
     /// </summary>
-    public abstract class StatsController : GameController<IStatsController, IStatsModule>, IStatsController
+    public class StatsController : GameController<IStatsController, IStatsModule>, IStatsController
     {
         /// <summary>
         /// Makes a new instance of the <see cref="StatController"/> class.
         /// </summary>
-        protected StatsController(IServiceLocator services)
+        public StatsController(IServiceLocator services)
             : base(new StatsModule(), services)
+        {
+            var campaigns = services.Get<ICampaigns>();
+            Health = new StatController(services, () => (int)((Strength.Current * 0.4) + (Vitality.Current * 0.6)));
+            Stamina = new StatController(services, () => (int)((Vitality.Current * 0.4) + (Agility.Current * 0.6)));
+            Mana = new StatController(services, () => Magic.Current > 0 ? (int)((Intelligence.Current * 0.4) + (Magic.Current * 0.6)) : 0);
+            Humans = new StatController(services, () => campaigns.Current.StatCap);
+            Sidhe = new StatController(services, () => campaigns.Current.StatCap);
+            Giants = new StatController(services, () => campaigns.Current.StatCap);
+            Strength = new LearningStatController(services);
+            Intelligence = new LearningStatController(services);
+            Agility = new LearningStatController(services);
+            Vitality = new LearningStatController(services);
+            Luck = new LearningStatController(services);
+            WeaponUse = new LearningStatController(services);
+            Parry = new LearningStatController(services);
+            Dodge = new LearningStatController(services);
+            Stealth = new LearningStatController(services);
+            LockPicking = new LearningStatController(services);
+            Throwing = new LearningStatController(services);
+            Climbing = new LearningStatController(services);
+            Magic = new LearningStatController(services);
+        }
+
+        public override int Width { get; }
+
+        public override int Height { get; }
+
+        public IStatController Health { get; }
+
+        public IStatController Stamina { get; }
+
+        public IStatController Mana { get; }
+
+        public IStatController Humans { get; }
+
+        public IStatController Sidhe { get; }
+
+        public IStatController Giants { get; }
+
+        public ILearningStatController Strength { get; }
+
+        public ILearningStatController Intelligence { get; }
+
+        public ILearningStatController Agility { get; }
+
+        public ILearningStatController Vitality { get; }
+
+        public ILearningStatController Luck { get; }
+
+        public ILearningStatController WeaponUse { get; }
+
+        public ILearningStatController Parry { get; }
+
+        public ILearningStatController Dodge { get; }
+
+        public ILearningStatController Stealth { get; }
+
+        public ILearningStatController LockPicking { get; }
+
+        public ILearningStatController Throwing { get; }
+
+        public ILearningStatController Climbing { get; }
+
+        public ILearningStatController Magic { get; }
+
+        public override bool Interact(int x, int y, Interaction interaction)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public override void Load()
         {
         }
 
-        /// <summary>
-        /// Gets the health stat.
-        /// </summary>
-        public abstract IStatController Health { get; }
+        public override void Unload()
+        {
+        }
 
-        /// <summary>
-        /// Gets the stamina stat.
-        /// </summary>
-        public abstract IStatController Stamina { get; }
+        public override void Update(GameTime time)
+        {
+        }
 
-        /// <summary>
-        /// Gets the mana stat.
-        /// </summary>
-        public abstract IStatController Mana { get; }
-
-        /// <summary>
-        /// Gets the humans stat.
-        /// </summary>
-        public abstract IStatController Humans { get; }
-
-        /// <summary>
-        /// Gets the sidhe stat.
-        /// </summary>
-        public abstract IStatController Sidhe { get; }
-
-        /// <summary>
-        /// Gets the giants stat.
-        /// </summary>
-        public abstract IStatController Giants { get; }
-
-        /// <summary>
-        /// Gets the strength stat.
-        /// </summary>
-        public abstract ILearningStatController Strength { get; }
-
-        /// <summary>
-        /// Gets the intelligence stat.
-        /// </summary>
-        public abstract ILearningStatController Intelligence { get; }
-
-        /// <summary>
-        /// Gets the agility stat.
-        /// </summary>
-        public abstract ILearningStatController Agility { get; }
-
-        /// <summary>
-        /// Gets the vitality stat.
-        /// </summary>
-        public abstract ILearningStatController Vitality { get; }
-
-        /// <summary>
-        /// Gets the luck stat.
-        /// </summary>
-        public abstract ILearningStatController Luck { get; }
-
-        /// <summary>
-        /// Gets the weapon use stat.
-        /// </summary>
-        public abstract ILearningStatController WeaponUse { get; }
-
-        /// <summary>
-        /// Gets the parry stat.
-        /// </summary>
-        public abstract ILearningStatController Parry { get; }
-
-        /// <summary>
-        /// Gets the dodge stat.
-        /// </summary>
-        public abstract ILearningStatController Dodge { get; }
-
-        /// <summary>
-        /// Gets the stealth stat.
-        /// </summary>
-        public abstract ILearningStatController Stealth { get; }
-
-        /// <summary>
-        /// Gets the lock picking.
-        /// </summary>
-        public abstract ILearningStatController LockPicking { get; }
-
-        /// <summary>
-        /// Gets the throwing stat.
-        /// </summary>
-        public abstract ILearningStatController Throwing { get; }
-
-        /// <summary>
-        /// Gets the climbing stat.
-        /// </summary>
-        public abstract ILearningStatController Climbing { get; }
-
-        /// <summary>
-        /// Gets the magic stat.
-        /// </summary>
-        public abstract ILearningStatController Magic { get; }
+        public override void Draw(GameTime time)
+        {
+        }
     }
 }
