+++ src/Hero6.Engine/ModuleController/Campaigns/ICampaignController.cs
@@ -24,41 +24,41 @@ namespace LateStartStudio.Hero6.ModuleController.Campaigns
         /// </summary>
         /// <typeparam name="T">The animation specified by type.</typeparam>
         /// <returns>The animation.</returns>
-        AnimationController GetAnimation<T>() where T : AnimationModule;
+        AnimationController GetAnimation<T>() where T : IAnimationModule;
 
         /// <summary>
         /// Gets any character animation from this campaign.
         /// </summary>
         /// <typeparam name="T">The character animation specified by type.</typeparam>
         /// <returns>The character animation.</returns>
-        CharacterAnimationController GetCharacterAnimation<T>() where T : CharacterAnimationModule;
+        CharacterAnimationController GetCharacterAnimation<T>() where T : ICharacterAnimationModule;
 
         /// <summary>
         /// Gets any character from this campaign.
         /// </summary>
         /// <typeparam name="T">The character specified by type.</typeparam>
         /// <returns>The character.</returns>
-        CharacterController GetCharacter<T>() where T : CharacterModule;
+        CharacterController GetCharacter<T>() where T : ICharacterModule;
 
         /// <summary>
         /// Gets any item from this campaign.
         /// </summary>
         /// <typeparam name="T">The item specified by type.</typeparam>
         /// <returns>The item.</returns>
-        ItemController GetItem<T>() where T : ItemModule;
+        ItemController GetItem<T>() where T : IItemModule;
 
         /// <summary>
         /// Gets any inventory item from this campaign.
         /// </summary>
         /// <typeparam name="T">The inventory item specified by type.</typeparam>
         /// <returns>The inventory item.</returns>
-        InventoryItemController GetInventoryItem<T>() where T : InventoryItemModule;
+        InventoryItemController GetInventoryItem<T>() where T : IInventoryItemModule;
 
         /// <summary>
         /// Gets any room from this campaign.
         /// </summary>
         /// <typeparam name="T">The room specified by type.</typeparam>
         /// <returns>The room.</returns>
-        RoomController GetRoom<T>() where T : RoomModule;
+        RoomController GetRoom<T>() where T : IRoomModule;
     }
 }
