+++ src/Hero6.Engine/ModuleController/Campaigns/Items/ItemController.cs
@@ -5,20 +5,84 @@
 // </copyright>
 
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace LateStartStudio.Hero6.ModuleController.Campaigns.Items
 {
     /// <summary>
     /// API for item controller.
     /// </summary>
-    public abstract class ItemController : GameController<IItemController, IItemModule>, IItemController
+    public class ItemController : GameController<IItemController, IItemModule>, IItemController
     {
+        private readonly ContentManager content;
+        private readonly SpriteBatch spriteBatch;
+
+        private Texture2D sprite;
+        private Vector2 position;
+
         /// <summary>
         /// Makes a new instance of the <see cref="ItemController"/> class.
         /// </summary>
         /// <param name="module">The module for this controller.</param>
-        protected ItemController(IItemModule module, IServiceLocator services) : base(module, services)
+        public ItemController(IItemModule module, IServiceLocator services) : base(module, services)
         {
+            content = services.Get<ContentManager>();
+            spriteBatch = services.Get<SpriteBatch>();
+        }
+
+        public override int Width => sprite.Width;
+
+        public override int Height => sprite.Height;
+
+        public override bool Interact(int x, int y, Interaction interaction)
+        {
+            if (Intersects(x, y))
+            {
+                switch (interaction)
+                {
+                    case Interaction.Eye:
+                        Module.Look?.Invoke();
+                        break;
+                    case Interaction.Mouth:
+                        Module.Talk?.Invoke();
+                        break;
+                    case Interaction.Hand:
+                        Module.Grab?.Invoke();
+                        break;
+                    case Interaction.Move:
+                    default:
+                        return false;
+                }
+
+                return true;
+            }
+
+            return false;
+        }
+
+        public override void Load()
+        {
+            sprite = content.Load<Texture2D>(Module.Sprite);
+        }
+
+        public override void Unload()
+        {
+        }
+
+        public override void Update(GameTime time)
+        {
+            position.X = X;
+            position.Y = Y;
+        }
+
+        public override void Draw(GameTime time)
+        {
+            if (Module.IsVisible)
+            {
+                spriteBatch.Draw(sprite, position, Color.White);
+            }
         }
     }
 }
