+++ src/Hero6.Engine/ModuleController/Campaigns/Rooms/Regions/HotspotsController.cs
@@ -4,28 +4,149 @@
 // 'LICENSE.CODE.md', which is a part of this source code package.
 // </copyright>
 
-using System.Drawing;
+using System.Collections.Generic;
+using LateStartStudio.Hero6.Extensions;
+using LateStartStudio.Hero6.ModuleController.Campaigns.Characters;
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
+using Color = System.Drawing.Color;
+using XnaColor = Microsoft.Xna.Framework.Color;
 
 namespace LateStartStudio.Hero6.ModuleController.Campaigns.Rooms.Regions
 {
     /// <summary>
     /// API for the hotspots controller.
     /// </summary>
-    public abstract class HotspotsController : GameController<IHotspotsController, IHotspotsModule>, IHotspotsController
+    public class HotspotsController : GameController<IHotspotsController, IHotspotsModule>, IHotspotsController
     {
+        private readonly string source;
+        private readonly ContentManager content;
+        private readonly Dictionary<Color, MonoGameHotspot> hotspots = new Dictionary<Color, MonoGameHotspot>();
+
+        private Texture2D texture;
+        private Color[,] buffer;
+
         /// <summary>
         /// Makes a new <see cref="HotspotsController"/> instance.
         /// </summary>
-        protected HotspotsController(IServiceLocator services) : base(new HotspotsModule(), services)
+        public HotspotsController(string source, IServiceLocator services) : base(new HotspotsModule(), services)
+        {
+            this.source = source;
+            content = services.Get<ContentManager>();
+        }
+
+        public override int Width { get; }
+
+        public override int Height { get; }
+
+        public Hotspot this[Color color] => hotspots[color];
+
+        public override bool Interact(int x, int y, Interaction interaction)
+        {
+            var pixel = buffer[y, x];
+            return pixel != Color.Transparent && hotspots[pixel].Interact(x, y, interaction);
+        }
+
+        public override void Load()
+        {
+            texture = content.Load<Texture2D>(source);
+            buffer = FindHotspots(texture);
+        }
+
+        public override void Unload()
         {
         }
 
-        /// <summary>
-        /// Gets the hotspot using the color value from the hotspots mask as a key.
-        /// </summary>
-        /// <param name="color">The color key in the hotspot mask.</param>
-        /// <returns>The hotspot from the mask with the color value.</returns>
-        public abstract Hotspot this[Color color] { get; }
+        public override void Update(GameTime time)
+        {
+        }
+
+        public override void Draw(GameTime time)
+        {
+        }
+
+        public void StandingOn(CharacterController character)
+        {
+            var pixel = buffer[character.Y, character.X];
+
+            if (hotspots.ContainsKey(pixel))
+            {
+                hotspots[pixel].StandingOn?.Invoke(new StandingOn(character.Module));
+            }
+        }
+
+        private static Color[,] CopyTextureData(Texture2D texture)
+        {
+            var result = new Color[texture.Height, texture.Width];
+            var data = new XnaColor[texture.Width * texture.Height];
+            texture.GetData(data);
+
+            for (var y = 0; y < texture.Height; y++)
+            {
+                for (var x = 0; x < texture.Width; x++)
+                {
+                    result[y, x] = data[(y * texture.Width) + x].ToDotNet();
+                }
+            }
+
+            return result;
+        }
+
+        private Color[,] FindHotspots(Texture2D texture)
+        {
+            var data = CopyTextureData(texture);
+
+            for (var y = 0; y < texture.Height; y++)
+            {
+                for (var x = 0; x < texture.Width; x++)
+                {
+                    var color = data[y, x];
+
+                    if (color != Color.Transparent)
+                    {
+                        if (!hotspots.ContainsKey(color))
+                        {
+                            hotspots[color] = new MonoGameHotspot();
+                        }
+
+                        hotspots[color].Add(new Point(x, y));
+                    }
+                }
+            }
+
+            return data;
+        }
+
+        private class MonoGameHotspot : Hotspot
+        {
+            private readonly List<Point> pixels = new List<Point>();
+
+            public void Add(Point pixel) => pixels.Add(pixel);
+
+            public bool Interact(int x, int y, Interaction interaction)
+            {
+                if (pixels.Contains(new Point(x, y)))
+                {
+                    switch (interaction)
+                    {
+                        case Interaction.Eye:
+                            Look?.Invoke();
+                            break;
+                        case Interaction.Hand:
+                            Grab?.Invoke();
+                            break;
+                        case Interaction.Mouth:
+                            Talk?.Invoke();
+                            break;
+                        default:
+                            return false;
+                    }
+                }
+
+                return true;
+            }
+        }
     }
 }
