+++ src/Hero6.Engine/ModuleController/UserInterfaces/Components/ButtonController.cs
@@ -4,14 +4,50 @@
 // 'LICENSE.CODE.md', which is a part of this source code package.
 // </copyright>
 
+using LateStartStudio.Hero6.Services.ControllerRepository;
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using Microsoft.Xna.Framework;
 
 namespace LateStartStudio.Hero6.ModuleController.UserInterfaces.Components
 {
-    public abstract class ButtonController : ComponentController<IButtonController, IButtonModule>, IButtonController
+    public class ButtonController : ComponentController<IButtonController, IButtonModule>, IButtonController
     {
-        protected ButtonController(IButtonModule module, IServiceLocator services) : base(module, services)
+        private readonly IControllerRepository controllerRepository;
+
+        public ButtonController(IButtonModule module, IServiceLocator services) : base(module, services)
         {
+            controllerRepository = services.Get<IControllerRepository>();
+        }
+
+        public override int Width => Module.Child.Width;
+
+        public override int Height => Module.Child.Height;
+
+        private IController ChildAsController => controllerRepository[Module.Child];
+
+        public override void Initialize()
+        {
+            ChildAsController.ToXnaGameLoop().Initialize();
+            base.Initialize();
+        }
+
+        public override void Load() => ChildAsController.ToXnaGameLoop().Load();
+
+        public override void Unload() => ChildAsController.ToXnaGameLoop().Unload();
+
+        public override void Update(GameTime time)
+        {
+            ChildAsController.ToXnaGameLoop().Update(time);
+            Module.Child.X = X;
+            Module.Child.Y = Y;
+        }
+
+        public override void Draw(GameTime time)
+        {
+            if (IsVisible)
+            {
+                ChildAsController.ToXnaGameLoop().Draw(time);
+            }
         }
     }
 }
