+++ src/Hero6.Engine/ModuleController/UserInterfaces/Components/ImageController.cs
@@ -5,16 +5,61 @@
 // </copyright>
 
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace LateStartStudio.Hero6.ModuleController.UserInterfaces.Components
 {
-    public abstract class ImageController : ComponentController<IImageController, IImageModule>, IImageController
+    public class ImageController : ComponentController<IImageController, IImageModule>, IImageController
     {
-        protected ImageController(IImageModule module, string source, IServiceLocator services) : base(module, services)
+        private readonly ContentManager content;
+        private readonly SpriteBatch spriteBatch;
+
+        private Texture2D image;
+        private Vector2 position;
+
+        public ImageController(IImageModule module, string source, IServiceLocator services) : base(module, services)
         {
             Source = source;
+            content = services.Get<ContentManager>();
+            spriteBatch = services.Get<SpriteBatch>();
         }
 
         public string Source { get; }
+
+        public override int X
+        {
+            get { return (int)position.X; }
+            set { position.X = value; }
+        }
+
+        public override int Y
+        {
+            get { return (int)position.Y; }
+            set { position.Y = value; }
+        }
+
+        public override int Width => image?.Width ?? 0;
+
+        public override int Height => image?.Height ?? 0;
+
+        public override void Load() => image = content.Load<Texture2D>(Source);
+
+        public override void Unload()
+        {
+        }
+
+        public override void Update(GameTime time)
+        {
+        }
+
+        public override void Draw(GameTime time)
+        {
+            if (IsVisible)
+            {
+                spriteBatch.Draw(image, position, Color.White);
+            }
+        }
     }
 }
