--- src/Hero6.Engine/ModuleController/UserInterfaces/Components/MonoGameButtonController.cs
@@ -1,53 +1,0 @@
-﻿// <copyright file="MonoGameButtonController.cs" company="Late Start Studio">
-// Copyright (C) Late Start Studio
-// This file is subject to the terms and conditions of the MIT license specified in the file
-// 'LICENSE.CODE.md', which is a part of this source code package.
-// </copyright>
-
-using LateStartStudio.Hero6.Services.ControllerRepository;
-using LateStartStudio.Hero6.Services.DependencyInjection;
-using Microsoft.Xna.Framework;
-
-namespace LateStartStudio.Hero6.ModuleController.UserInterfaces.Components
-{
-    public class MonoGameButtonController : ButtonController
-    {
-        private readonly IControllerRepository controllerRepository;
-
-        public MonoGameButtonController(IButtonModule module, IServiceLocator services) : base(module, services)
-        {
-            controllerRepository = services.Get<IControllerRepository>();
-        }
-
-        public override int Width => Module.Child.Width;
-
-        public override int Height => Module.Child.Height;
-
-        private IController ChildAsController => controllerRepository[Module.Child];
-
-        public override void Initialize()
-        {
-            ChildAsController.ToXnaGameLoop().Initialize();
-            base.Initialize();
-        }
-
-        public override void Load() => ChildAsController.ToXnaGameLoop().Load();
-
-        public override void Unload() => ChildAsController.ToXnaGameLoop().Unload();
-
-        public override void Update(GameTime time)
-        {
-            ChildAsController.ToXnaGameLoop().Update(time);
-            Module.Child.X = X;
-            Module.Child.Y = Y;
-        }
-
-        public override void Draw(GameTime time)
-        {
-            if (IsVisible)
-            {
-                ChildAsController.ToXnaGameLoop().Draw(time);
-            }
-        }
-    }
-}
