+++ src/Hero6.Engine/ModuleController/UserInterfaces/Input/Mouse/CursorController.cs
@@ -5,14 +5,45 @@
 // </copyright>
 
 using LateStartStudio.Hero6.Services.DependencyInjection;
+using LateStartStudio.Hero6.Services.UserInterfaces.Input.Mouse;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace LateStartStudio.Hero6.ModuleController.UserInterfaces.Input.Mouse
 {
-    public abstract class CursorController : Controller<ICursorController, ICursorModule>, ICursorController
+    public class CursorController : Controller<ICursorController, ICursorModule>, ICursorController
     {
-        protected CursorController(ICursorModule module, IServiceLocator services) : base(module, services)
+        private readonly ContentManager content;
+        private readonly SpriteBatch spriteBatch;
+        private readonly IMouse mouse;
+
+        private Texture2D cursor;
+
+        public CursorController(ICursorModule module, IServiceLocator services) : base(module, services)
+        {
+            content = services.Get<ContentManager>();
+            spriteBatch = services.Get<SpriteBatch>();
+            mouse = services.Get<IMouse>();
+        }
+
+        public override int Width => cursor.Width;
+
+        public override int Height => cursor.Height;
+
+        public override void Load() => cursor = content.Load<Texture2D>(Module.Source);
+
+        public override void Unload()
         {
         }
+
+        public override void Update(GameTime time)
+        {
+            X = mouse.X;
+            Y = mouse.Y;
+        }
+
+        public override void Draw(GameTime time) => spriteBatch.Draw(cursor, new Vector2(X, Y), Color.White);
 
         public bool Equals<T>() => typeof(T) == Module.GetType();
     }
