+++ src/Hero6.Tests.HelperTools/ModuleController/Campaigns/Rooms/RoomTestBase.cs
@@ -4,7 +4,6 @@
 // 'LICENSE.CODE.md', which is a part of this source code package.
 // </copyright>
 
-using LateStartStudio.Hero6.ModuleController.Campaigns.Rooms.Regions;
 using NSubstitute;
 
 namespace LateStartStudio.Hero6.ModuleController.Campaigns.Rooms
@@ -12,14 +11,6 @@ namespace LateStartStudio.Hero6.ModuleController.Campaigns.Rooms
     public abstract class RoomTestBase<TModule> : ModuleControllerTestBase<TModule, IRoomController>
         where TModule : IRoomModule
     {
-        protected override IRoomController MakeController()
-        {
-            var controller = Substitute.For<RoomController>(Module, Services.Services);
-            var hotspotsController = Substitute.For<HotspotsController>(Services.Services);
-            hotspotsController.PreInitialize();
-            hotspotsController.Initialize();
-            controller.Hotspots.Returns(hotspotsController);
-            return controller;
-        }
+        protected override IRoomController MakeController() => Substitute.For<IRoomController>();
     }
 }
